Hydropod app data view for ph, water elvel, water flow and temperature

+++ hydropod_hydroponics/lib/models/models.dart
@@ -1,6 +1,6 @@
 /// Institution model
 class DeviceData {
-  int ph;
+  double ph;
   int waterFLow;
   int totalWaterFlow;
   int waterLevel;
